cxr-main-app/requirements.txt
Flask
requests==2.27.1
Pillow==9.0.1
numpy==1.22.3
Werkzeug
